Update the Logging.Entry properties
Resolution and EventId properties were not updated when the
Logging.Entry D-Bus object is created. So code that relies on D-Bus
signal to detect new logging entries will need to make additional
D-Bus calls to get all the properties.

This patch delays sending the D-Bus signal after processing the
FQPN.

Signed-off-by: Tom Joseph <rushtotom@gmail.com>