Fix id for SEL entry
The ID for SEL entries currently is from 1 to maximum of uint32 which
is not consistent with the ID from ipmitool and not following the ipmi
spec. The ID for SEL should be uint16 type with 0x0 and 0xffff
reserved.
This patch is to wrap the ID if reaching the maximum of uint16.

Tested:
Generated more than 65534 SEL entries, the SEL ID will be wrapped and
the maximum ID for SEL entry is 65534(0xfffe).

Fixes jira https://redmine.mellanox.com/issues/4126832

Signed-off-by: Sean Zhang <xiazhang@nvidia.com>