Rsyslog dynamic filter handling
Issue we have seen is by default we are
adding the all in the config due to which
filter is allowing everything.

Fix we have done is handled the all filter
in the default case.

Tested by pushing the kern log in daemon
server setting and daemon log in kern
server setting and all logs setting.

Fixes nvbug https://nvbugspro.nvidia.com/bug/5045481
Signed-off-by: Deepak Kamat <dkamat@nvidia.com>