phosphor-logging: introduce a new key for max logs as single limit
Introduces a JSON option to set the single capacity limit for
informational (and lower severity) log entries, controlled by the
DefaultCapacity JSON option. When it is not set, informational and
error entries are tracked and capped independently. When it is set,
only default capacity is used for all log severities.

Tested:
1. Set DefaultCapacity to 30 entries in
   phosphor-logging-namespace.json.
3. Add entries with different severity
4. All the entries can be stored and the max limit is 30.

Fixes jira https://redmine.mellanox.com/issues/4127100

Signed-off-by: Qian Sun <qians@nvidia.com>